+++ pkg/clusters/addons/metallb/metallb.go
@@ -144,7 +144,7 @@ func (a *addon) DumpDiagnostics(ctx context.Context, cluster clusters.Cluster) (
 var (
 	defaultStartIP = net.ParseIP("0.0.0.100")
 	defaultEndIP   = net.ParseIP("0.0.0.250")
-	metalManifest  = "https://github.com/metallb/metallb/config/native?ref=v0.13.7?timeout=2m"
+	metalManifest  = "https://github.com/metallb/metallb/config/native?ref=v0.13.7&timeout=2m"
 	secretKeyLen   = 128
 )