+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.25.0
+## Unreleased 
 
 - GKE cluster builder allows creating a subnet for the cluster instead of using 
   a default one.