+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - GKE cluster is able to wait for its cleanup synchronously. 
   [#491](https://github.com/Kong/kubernetes-testing-framework/pull/491)
 
+- MetalLB addon will use an extended timeout when fetching manifests from GH which 
+  should improve its stability. 
+  [#492](https://github.com/Kong/kubernetes-testing-framework/pull/492)
+
 ## v0.24.1
 
 - Golang dependencies for several Kubernetes libraries were updated to